Found a way to make a 16x16 grid using flexbox.
We can adjust the number of boxes in the grid with the for loop, in this case, 19x19 = 361 squares

Each time the number of the area of the grid is increased by one, we have to remove 6 pixels from the flex basis of the divs, in order to
mantain the grid uniform.
This "6" value might change if i remove the pixel of the border, but it can be easily solved later when we don't need a visual reference.